Modify directory structure and initialization page
