+++ Soundbyte-App/src/Initialization/index.html
@@ -7,14 +7,14 @@
 
   <link rel="preconnect" href="https://fonts.gstatic.com"> 
   <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@200;400;500;700&display=swap" rel="stylesheet">
-  <link rel="stylesheet" type="text/css" href="./_index.css" />
-  <link rel="stylesheet" type="text/css" href="./setupSongLibrary.css" />
+  <link rel="stylesheet" type="text/css" href="../_index.css" />
+  <link rel="stylesheet" type="text/css" href="./index.css" />
 
 </head>
 <body>
   <div id=container>
     <nav>            
-        <img src="./img/SoundByte-logo___.png" class="logo" alt="site log" />
+        <img src="../img/SoundByte-logo___.png" class="logo" alt="site log" />
         
         <ul>
             
@@ -28,7 +28,7 @@
             <p>Welcome,</p>
             <p>It looks like we need to do the first time set-up.<p/>
             <div class="fileUpload">
-            Please provide your own JSON file.
+            Please provide your own JSON directory.
             <form>
         
             <label class="file">
@@ -51,6 +51,6 @@
     </div>
 </div>
 
-  <script src="setupSongLibrary.js"></script>
+  <script src="./index.js"></script>
 </body>
 </html>
