--- Soundbyte-App/src/Initialization/initialization.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Hello World!</title>
-    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';" /> 
-    <script src="./main/loadLibrary.js" type="text/javascript"></script> 
-</head>
-<body style="background: white;" onload="loadLibrary()">
-
-</body>
-</html>
